requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiohttp-retry==2.8.3
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
frozenlist==1.4.1
idna==3.8
multidict==6.0.5
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
twilio==9.2.4
urllib3==2.2.2
yarl==1.9.7